Fix Clang warning ("operands of ? are integers of different signs")
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9214 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/arch/x86_x64/cache.cpp
@@ -350,12 +350,12 @@ struct Characteristics	// POD
 
 	size_t NumEntries() const
 	{
-		return IsTLB()? smallSize : (flags & ~0xF);
+		return IsTLB()? (size_t)smallSize : (flags & ~0xF);
 	}
 
 	size_t EntrySize() const
 	{
-		return IsTLB()? (flags & ~0xF) : -smallSize;
+		return IsTLB()? (flags & ~0xF) : (size_t)(-smallSize);
 	}
 
 	u8 descriptor;
